various fixes
