+++ pyleoclim/core/__init__.py
@@ -19,5 +19,4 @@ from .multivardecomp import MultivariateDecomp
 from .ssares import SsaRes
 from .lipd import Lipd
 from .lipdseries import LipdSeries
-from .resolution import Resolution
-from .multipleresolution import MultipleResolution
+from .resolutions import Resolution, MultipleResolution
